--- BDD/WebAPI_Etienne/appsettings.json
@@ -1,18 +1,0 @@
-{
-  "ConnectionStrings": {
-    "MurVegetalDb": "mongodb://localhost:27017"
-  },
-  "Logging": {
-    "IncludeScopes": false,
-    "Debug": {
-      "LogLevel": {
-        "Default": "Warning"
-      }
-    },
-    "Console": {
-      "LogLevel": {
-        "Default": "Warning"
-      }
-    }
-  }
-}
